classify: resort to composite classifier in the end
If not everything is classified and there are still more than one classifier
left we should try to classify by composition to detect situations where, for
example, we have 2 classifiers and 4 implementations, the first classifier
gives the same results for 1 and 2, and also for 3 and 4. The second one is able
to detect 1&3 vs 2&4. While none of classifiers can precisely detect any of the
implementations, it's clear that two of them can distinguish between all of
them.

In fact, we could start immediately with checking the composite classifier but
we want more information on what is able to detect what in the output, so we
first try to find classifiers that are able to precisely detect one of
implementations.